requirements.txt
python-telegram-bot>=11.1.0
simplesqlitewrap>=0.1.14
pillow
pyyaml
toml
emoji
